move gravity calculations to own file
